moving slot related functions to util.js file
