+++ lambda/custom/helperFunctions.js
@@ -529,57 +529,6 @@ function emojiTranslateFunc(str) {
 	return emojiTranslate.translate(str, onlyEmoji);
 }
 
-function slotValue(slot) {
-	let { value } = slot;
-	const resolution = (slot.resolutions && slot.resolutions.resolutionsPerAuthority && slot.resolutions.resolutionsPerAuthority.length > 0) ? slot.resolutions.resolutionsPerAuthority[0] : null;
-	if (resolution && resolution.status.code === 'ER_SUCCESS_MATCH') {
-		const resolutionValue = resolution.values[0].value;
-		value = resolutionValue.name;
-	}
-	return value;
-}
-
-function getStaticAndDynamicSlotValuesFromSlot(slot) {
-
-	const result = {
-		name: slot.name,
-		value: slot.value,
-	};
-
-	if (((slot.resolutions || {}).resolutionsPerAuthority || [])[0] || {}) {
-		slot.resolutions.resolutionsPerAuthority.forEach((authority) => {
-			const slotValue = {
-				authority: authority.authority,
-				statusCode: authority.status.code,
-				synonym: slot.value || undefined,
-				resolvedValues: slot.value,
-			};
-			if (authority.values && authority.values.length > 0) {
-				slotValue.resolvedValues = [];
-
-				authority.values.forEach((value) => {
-					slotValue.resolvedValues.push(value);
-				});
-
-			}
-
-			if (authority.authority.includes('amzn1.er-authority.echo-sdk.dynamic')) {
-				result.dynamic = slotValue;
-			} else {
-				result.static = slotValue;
-			}
-		});
-	}
-
-	if (result.hasOwnProperty('dynamic') && result.dynamic.statusCode === 'ER_SUCCESS_MATCH') {
-		return result.dynamic.resolvedValues[0].value.name;
-	} else if (result.hasOwnProperty('static') && result.static.statusCode === 'ER_SUCCESS_MATCH') {
-		return result.static.resolvedValues[0].value.name;
-	} else {
-		return result.value;
-	}
-}
-
 const createGroup = async (channelName, headers) =>
 	await axios
 		.post(
@@ -1012,8 +961,6 @@ module.exports.replaceWhitespacesFunc = replaceWhitespacesFunc;
 module.exports.replaceWhitespacesDots = replaceWhitespacesDots;
 module.exports.emojiTranslateFunc = emojiTranslateFunc;
 module.exports.readMessages = readMessages;
-module.exports.slotValue = slotValue;
-module.exports.getStaticAndDynamicSlotValuesFromSlot = getStaticAndDynamicSlotValuesFromSlot;
 module.exports.createGroup = createGroup;
 module.exports.deleteGroup = deleteGroup;
 module.exports.getGroupId = getGroupId;
